upgrades to the machine crafting and state validation to handle module math
* building/building.go: added simplified crafting emulation
* building/inventory.go: better implement slot constraints
* constants/buildings.go: hardcode vanilla buildings (for state validation)
* data/data.go: add Tool (for science packs)
* data/gen_get.bash: add GetTool
* state/state.go: functions to construct/mine buildings
* tas/tas.go: better state validation
* tas/task.go: MachineCraft accounts for batching, inventory limitations, and modules

+++ calc/recipe.go
@@ -358,7 +358,7 @@ func Handcraft(inventory Items[uint], recipe *data.Recipe, amount uint) (newInve
 // OneStackRecipe returns the number of crafts of a recipe that can be done if given one
 // stack of each input item, and at most produces one stack of output.
 // Fluid inputs are skipped since they're not subject to the same stacking constraints
-func OneStackRecipe(r *data.Recipe) int {
+func OneStackRecipe(r *data.Recipe) uint {
 	count := math.MaxInt
 
 	for _, ing := range r.Ingredients {
@@ -378,5 +378,5 @@ func OneStackRecipe(r *data.Recipe) int {
 		count = c
 	}
 
-	return count
+	return uint(count)
 }
